requirements.txt
MetaTrader5
notion-client
python-dotenv
pytz
requests
psutil
